requirements.txt
# Core dependencies
boto3>=1.28.0
botocore>=1.31.0

# Data processing
pyarrow>=14.0.0
pyspark>=3.5.0          # pre-installed on Databricks/Glue
delta-spark>=3.0.0      # Delta Lake for local Spark testing

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
